docker_compose/etl/requirements.txt
# jupyterlab
flake8==4.0.1
python-dotenv==0.20.0
psycopg2==2.9.3
pydantic
python-dateutil
# memory-profiler
isort
loguru